add refactor type for quantities

+++ include/exchange-disruptor/matching_engine/command/limit_command.h
@@ -5,15 +5,15 @@
 
 class LimitCommand : public Command {
 public:
-    LimitCommand(uint64_t clientId, uint64_t orderId, std::string symbol, Side side, int64_t price, double qty);
+    LimitCommand(uint64_t clientId, uint64_t orderId, std::string symbol, Side side, int64_t price, int64_t qty);
     CommandType getType() const override;
     int64_t getPrice() const;
-    double getQty() const;
+    int64_t getQty() const;
     Side getSide() const;
 
 private:
     int64_t _price;
-    double _qty;
+    int64_t _qty;
     Side _side;
 };
