+++ include/exchange-disruptor/matching_engine/event/fill_event.h
@@ -5,15 +5,15 @@
 
 class FillEvent : public Event {
 public:
-    FillEvent(uint64_t clientId, uint64_t orderId, uint64_t matchId, int64_t price, double qty);
+    FillEvent(uint64_t clientId, uint64_t orderId, uint64_t matchId, int64_t price, int64_t qty);
     EventType getType() const override;
     uint64_t getMatchId() const;
     int64_t getPrice() const;
-    double getQty() const;
+    int64_t getQty() const;
 private:
     uint64_t _matchId;
     int64_t _price;
-    double _qty;
+    int64_t _qty;
 };
 
 #endif // FILL_EVENT_H
